backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
prophet==1.1.5
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
requests==2.31.0
reportlab==4.0.7
pillow==10.1.0
gunicorn==21.2.0